* 'develop' of github.com:ampproject/amp-wp: (643 commits)
  Update CODEOWNERS
  Update amphtml spec to 2111242025000
  Use CF7 instead of Autoptimize for AMP-incompatible plugin in E2E tests
  Replace autoptimize with akismet for E2E tests
  Pin NodeJS to v14 to match WP core and Gutenberg
  Regenerate package-lock.json at lockfileVersion 1
  Update Gutenberg package dependencies
  Stop mapping E2E test plugins dir to container volume
  Add demo plugin ZIP archive
  Fix script command
  Fix lint issues
  Ensure React state updates are not performed on unmounted components
  Use different Jest configs for E2E tests; use `.env` config file
  Refactor "Other Settings" E2E tests
  Ensure selector is available before scrolling element into view
  Use more robust internal activate/deactivate functions in E2E tests
  Introduce general `saveSettings` util function for E2E tests
  Remove unnecessary cleanup functions in E2E tests
  Install Hestia theme globally as it is used in multiple tests
  Use local demo plugin in E2E tests
  ...

+++ includes/uninstall-functions.php
@@ -19,6 +19,7 @@ function delete_options() {
 	delete_option( 'amp-options' );
 	delete_option( 'amp_css_transient_monitor_time_series' );
 	delete_option( 'amp_customize_setting_modified_timestamps' );
+	delete_option( 'amp_url_validation_queue' ); // See Validation\URLValidationCron::OPTION_KEY.
 
 	$theme_mod_name = 'amp_customize_setting_modified_timestamps';
 	remove_theme_mod( $theme_mod_name );
@@ -175,7 +176,7 @@ function delete_transients() {
 
 	// phpcs:ignore WordPress.DB.DirectDatabaseQuery.DirectQuery, WordPress.DB.DirectDatabaseQuery.NoCaching -- Cannot cache result since we're deleting the records.
 	$wpdb->query(
-	// phpcs:ignore WordPress.DB.PreparedSQL.NotPrepared -- See use of prepare in foreach loop above.
+		// phpcs:ignore WordPress.DB.PreparedSQL.NotPrepared -- See use of prepare in foreach loop above.
 		"DELETE FROM $wpdb->options WHERE " . implode( ' OR ', $where_clause )
 	);
 }
@@ -187,10 +188,17 @@ function delete_transients() {
  * @internal
  */
 function remove_plugin_data() {
+	$options = get_option( 'amp-options' );
 
-	delete_options();
-	delete_user_metadata();
-	delete_posts();
-	delete_terms();
-	delete_transients();
+	if (
+		is_array( $options ) && array_key_exists( 'delete_data_at_uninstall', $options )
+			? $options['delete_data_at_uninstall']
+			: true
+	) {
+		delete_options();
+		delete_user_metadata();
+		delete_posts();
+		delete_terms();
+		delete_transients();
+	}
 }
